Move client sources to categorized subdirectories
