--- data/points.js
@@ -1,487 +1,0 @@
-var points = [{"lat": 44.7559, "lon": 33.85574}, {"lat": 45.05836, "lon": 34.59474}, {
-  "lat": 45.71102,
-  "lon": 34.41141
-}, {"lat": 45.54417, "lon": 33.07932}, {"lat": 45.18127, "lon": 34.94802}, {
-  "lat": 45.95272,
-  "lon": 33.79377
-}, {"lat": 45.29543, "lon": 35.76771}, {"lat": 45.4422, "lon": 34.73894}, {
-  "lat": 45.11536,
-  "lon": 34.06225
-}, {"lat": 45.779, "lon": 33.48598}, {"lat": 45.13458, "lon": 33.60889}, {
-  "lat": 44.95829,
-  "lon": 34.11014
-}, {"lat": 45.33628, "lon": 34.93481}, {"lat": 45.50451, "lon": 32.69754}, {
-  "lat": 49.07421,
-  "lon": 27.67319
-}, {"lat": 48.36611, "lon": 29.5182}, {"lat": 49.23523, "lon": 28.48463}, {
-  "lat": 48.81054,
-  "lon": 29.39375
-}, {"lat": 49.0325, "lon": 28.11659}, {"lat": 49.10837, "lon": 29.20784}, {
-  "lat": 49.69597,
-  "lon": 28.82298
-}, {"lat": 48.21693, "lon": 28.65543}, {"lat": 48.38362, "lon": 28.86503}, {
-  "lat": 49.22733,
-  "lon": 29.0609
-}, {"lat": 49.32106, "lon": 28.09118}, {"lat": 48.44919, "lon": 27.80331}, {
-  "lat": 48.7084,
-  "lon": 27.53414
-}, {"lat": 48.97722, "lon": 28.84289}, {"lat": 49.18974, "lon": 29.517}, {
-  "lat": 48.20354,
-  "lon": 28.89422
-}, {"lat": 49.48517, "lon": 29.25848}, {"lat": 48.66274, "lon": 29.74634}, {
-  "lat": 48.53742,
-  "lon": 28.51467
-}, {"lat": 48.36302, "lon": 28.2316}, {"lat": 48.67441, "lon": 28.87499}, {
-  "lat": 49.01567,
-  "lon": 28.50969
-}, {"lat": 49.55915, "lon": 27.9542}, {"lat": 48.53794, "lon": 28.12878}, {
-  "lat": 48.21092,
-  "lon": 29.36302
-}, {"lat": 48.73861, "lon": 28.08312}, {"lat": 48.24903, "lon": 28.29186}, {
-  "lat": 50.8497,
-  "lon": 24.30777
-}, {"lat": 50.50312, "lon": 24.7716}, {"lat": 50.63787, "lon": 24.36716}, {
-  "lat": 51.62527,
-  "lon": 24.963
-}, {"lat": 50.84301, "lon": 25.4488}, {"lat": 51.21843, "lon": 24.69469}, {
-  "lat": 50.734,
-  "lon": 24.648
-}, {"lat": 50.74293, "lon": 25.31216}, {"lat": 51.75453, "lon": 25.49172}, {
-  "lat": 51.22753,
-  "lon": 24.03225
-}, {"lat": 51.29087, "lon": 25.51935}, {"lat": 51.67145, "lon": 24.52475}, {
-  "lat": 50.91648,
-  "lon": 25.27165
-}, {"lat": 51.43988, "lon": 24.4372}, {"lat": 51.08299, "lon": 24.52372}, {
-  "lat": 51.49069,
-  "lon": 23.92136
-}, {"lat": 47.66298, "lon": 33.72013}, {"lat": 48.2147, "lon": 36.02503}, {
-  "lat": 48.62841,
-  "lon": 34.33519
-}, {"lat": 48.46928, "lon": 35.01171}, {"lat": 47.88803, "lon": 33.39346}, {
-  "lat": 47.9845,
-  "lon": 34.55801
-}, {"lat": 48.92435, "lon": 34.90614}, {"lat": 48.25109, "lon": 36.74034}, {
-  "lat": 47.58282,
-  "lon": 34.38343
-}, {"lat": 48.63613, "lon": 35.25839}, {"lat": 48.52352, "lon": 35.86933}, {
-  "lat": 47.99171,
-  "lon": 34.44214
-}, {"lat": 47.84855, "lon": 33.98826}, {"lat": 47.55381, "lon": 34.15066}, {
-  "lat": 48.42447,
-  "lon": 33.72133
-}, {"lat": 48.31959, "lon": 35.52893}, {"lat": 48.20869, "lon": 34.87181}, {
-  "lat": 48.05076,
-  "lon": 33.86913
-}, {"lat": 48.11153, "lon": 34.3376}, {"lat": 48.94564, "lon": 34.49158}, {
-  "lat": 47.69783,
-  "lon": 33.27089
-}, {"lat": 49.02597, "lon": 34.70066}, {"lat": 47.91635, "lon": 37.57307}, {
-  "lat": 47.78744,
-  "lon": 38.46399
-}, {"lat": 48.3616, "lon": 38.0024}, {"lat": 47.83687, "lon": 36.83441}, {
-  "lat": 47.59432,
-  "lon": 37.50234
-}, {"lat": 47.1237, "lon": 37.1931}, {"lat": 48.44919, "lon": 37.09276}, {
-  "lat": 47.86949,
-  "lon": 37.40553
-}, {"lat": 48.26843, "lon": 37.17773}, {"lat": 48.40971, "lon": 36.86153}, {
-  "lat": 47.94176,
-  "lon": 37.50406
-}, {"lat": 47.11367, "lon": 38.08239}, {"lat": 47.05685, "lon": 37.30682}, {
-  "lat": 48.84367,
-  "lon": 37.64568
-}, {"lat": 47.74984, "lon": 38.02454}, {"lat": 47.41081, "lon": 38.01647}, {
-  "lat": 48.0451,
-  "lon": 38.45129
-}, {"lat": 48.11565, "lon": 37.88584}, {"lat": 50.02075, "lon": 29.01781}, {
-  "lat": 50.29644,
-  "lon": 27.67113
-}, {"lat": 49.89475, "lon": 28.57905}, {"lat": 50.28168, "lon": 29.51872}, {
-  "lat": 50.3551,
-  "lon": 28.2621
-}, {"lat": 50.14864, "lon": 27.93085}, {"lat": 50.87185, "lon": 27.80674}, {
-  "lat": 50.27447,
-  "lon": 28.65749
-}, {"lat": 50.94807, "lon": 28.64101}, {"lat": 50.31326, "lon": 29.05884}, {
-  "lat": 51.08299,
-  "lon": 28.39159
-}, {"lat": 49.92256, "lon": 27.76125}, {"lat": 50.76902, "lon": 29.242}, {
-  "lat": 51.20041,
-  "lon": 29.09077
-}, {"lat": 50.5874, "lon": 27.62426}, {"lat": 51.31319, "lon": 28.79534}, {
-  "lat": 51.22221,
-  "lon": 27.64984
-}, {"lat": 49.9435, "lon": 29.47872}, {"lat": 50.50381, "lon": 29.24612}, {
-  "lat": 49.72206,
-  "lon": 29.21831
-}, {"lat": 50.46707, "lon": 28.27555}, {"lat": 50.45299, "lon": 28.66556}, {
-  "lat": 50.05234,
-  "lon": 28.11556
-}, {"lat": 48.20543, "lon": 22.64591}, {"lat": 48.88967, "lon": 22.45829}, {
-  "lat": 48.14072,
-  "lon": 23.04674
-}, {"lat": 48.71681, "lon": 23.18029}, {"lat": 48.31255, "lon": 23.04108}, {
-  "lat": 48.53176,
-  "lon": 23.49838
-}, {"lat": 48.44147, "lon": 22.72162}, {"lat": 48.73741, "lon": 22.48884}, {
-  "lat": 48.05574,
-  "lon": 24.20357
-}, {"lat": 48.54618, "lon": 22.99027}, {"lat": 48.01163, "lon": 23.57838}, {
-  "lat": 48.6212,
-  "lon": 22.2971
-}, {"lat": 48.17814, "lon": 23.29067}, {"lat": 46.69618, "lon": 35.15385}, {
-  "lat": 46.76365,
-  "lon": 36.7987
-}, {"lat": 47.42849, "lon": 35.24877}, {"lat": 47.27743, "lon": 34.7168}, {
-  "lat": 47.01994,
-  "lon": 34.92485
-}, {"lat": 47.94313, "lon": 35.43228}, {"lat": 47.66539, "lon": 36.25265}, {
-  "lat": 46.96157,
-  "lon": 35.04072
-}, {"lat": 47.48978, "lon": 34.41072}, {"lat": 46.78081, "lon": 34.90082}, {
-  "lat": 46.84673,
-  "lon": 35.37563
-}, {"lat": 47.27228, "lon": 35.21959}, {"lat": 46.88553, "lon": 35.21856}, {
-  "lat": 47.56222,
-  "lon": 35.77492
-}, {"lat": 47.48325, "lon": 36.25419}, {"lat": 46.72966, "lon": 35.64531}, {
-  "lat": 46.72863,
-  "lon": 36.3541
-}, {"lat": 46.50684, "lon": 34.96124}, {"lat": 47.25374, "lon": 35.69544}, {
-  "lat": 47.19383,
-  "lon": 36.20819
-}, {"lat": 48.80659, "lon": 24.53676}, {"lat": 48.15531, "lon": 24.81812}, {
-  "lat": 49.12451,
-  "lon": 24.72834
-}, {"lat": 48.67613, "lon": 25.49154}, {"lat": 48.97035, "lon": 24.0156}, {
-  "lat": 49.02048,
-  "lon": 24.36562
-}, {"lat": 48.52489, "lon": 25.03733}, {"lat": 48.32079, "lon": 25.1005}, {
-  "lat": 48.63407,
-  "lon": 24.56749
-}, {"lat": 49.41049, "lon": 24.60955}, {"lat": 48.93705, "lon": 24.15379}, {
-  "lat": 48.45571,
-  "lon": 25.54562
-}, {"lat": 48.8665, "lon": 25.00076}, {"lat": 48.90341, "lon": 24.84678}, {
-  "lat": 50.36407,
-  "lon": 31.32958
-}, {"lat": 49.80617, "lon": 30.10443}, {"lat": 49.54731, "lon": 30.8733}, {
-  "lat": 50.34845,
-  "lon": 30.95553
-}, {"lat": 50.64011, "lon": 29.918}, {"lat": 50.49917, "lon": 30.77236}, {
-  "lat": 50.17456,
-  "lon": 30.32209
-}, {"lat": 49.52276, "lon": 29.92745}, {"lat": 50.58758, "lon": 30.49496}, {
-  "lat": 50.49419,
-  "lon": 31.77023
-}, {"lat": 50.31669, "lon": 31.06281}, {"lat": 49.8654, "lon": 30.823}, {"lat": 50.27, "lon": 30.3125}, {
-  "lat": 50.4669,
-  "lon": 29.81226
-}, {"lat": 49.64395, "lon": 30.96136}, {"lat": 50.11087, "lon": 30.62714}, {
-  "lat": 50.06504,
-  "lon": 31.44425
-}, {"lat": 51.07664, "lon": 29.63768}, {"lat": 49.68498, "lon": 30.47024}, {
-  "lat": 49.73322,
-  "lon": 29.66343
-}, {"lat": 49.38492, "lon": 30.19232}, {"lat": 49.55521, "lon": 30.50251}, {
-  "lat": 49.3705,
-  "lon": 29.68214
-}, {"lat": 50.07963, "lon": 29.91285}, {"lat": 50.28322, "lon": 31.76148}, {
-  "lat": 48.67596,
-  "lon": 33.11623
-}, {"lat": 48.34139, "lon": 30.52551}, {"lat": 48.06141, "lon": 32.15973}, {
-  "lat": 48.33692,
-  "lon": 29.86633
-}, {"lat": 48.38121, "lon": 30.46835}, {"lat": 48.38739, "lon": 31.17731}, {
-  "lat": 48.11445,
-  "lon": 32.75917
-}, {"lat": 48.72093, "lon": 32.66956}, {"lat": 48.50344, "lon": 32.26015}, {
-  "lat": 48.24938,
-  "lon": 32.21277
-}, {"lat": 48.63905, "lon": 31.63582}, {"lat": 48.36525, "lon": 32.65634}, {
-  "lat": 48.64832,
-  "lon": 30.8139
-}, {"lat": 48.79389, "lon": 31.65024}, {"lat": 48.32268, "lon": 31.52252}, {
-  "lat": 48.2341,
-  "lon": 30.8733
-}, {"lat": 48.90427, "lon": 33.4477}, {"lat": 48.71046, "lon": 32.13741}, {
-  "lat": 49.04554,
-  "lon": 33.20738
-}, {"lat": 48.31255, "lon": 30.22888}, {"lat": 47.96408, "lon": 32.53704}, {
-  "lat": 48.11651,
-  "lon": 39.08884
-}, {"lat": 49.20416, "lon": 39.5767}, {"lat": 49.52722, "lon": 38.73659}, {
-  "lat": 48.30774,
-  "lon": 39.72021
-}, {"lat": 49.05361, "lon": 38.22521}, {"lat": 48.40267, "lon": 39.20797}, {
-  "lat": 49.52534,
-  "lon": 39.56228
-}, {"lat": 49.38286, "lon": 40.12756}, {"lat": 48.96932, "lon": 39.01125}, {
-  "lat": 49.5377,
-  "lon": 39.10154
-}, {"lat": 48.44009, "lon": 38.79976}, {"lat": 48.63029, "lon": 38.36632}, {
-  "lat": 49.40843,
-  "lon": 38.15054
-}, {"lat": 48.07325, "lon": 39.64039}, {"lat": 48.69552, "lon": 38.98344}, {
-  "lat": 48.4014,
-  "lon": 39.282
-}, {"lat": 49.2917, "lon": 38.93074}, {"lat": 48.52404, "lon": 38.37095}, {
-  "lat": 50.06967,
-  "lon": 25.16573
-}, {"lat": 49.97234, "lon": 24.61504}, {"lat": 49.78695, "lon": 23.67159}, {
-  "lat": 49.35316,
-  "lon": 23.51006
-}, {"lat": 49.38303, "lon": 24.14194}, {"lat": 49.79433, "lon": 24.89382}, {
-  "lat": 50.10898,
-  "lon": 24.34467
-}, {"lat": 49.79193, "lon": 23.14837}, {"lat": 50.05646, "lon": 23.97303}, {
-  "lat": 49.52619,
-  "lon": 23.98092
-}, {"lat": 49.66816, "lon": 24.55856}, {"lat": 49.72103, "lon": 23.90677}, {
-  "lat": 50.27962,
-  "lon": 24.64285
-}, {"lat": 49.51658, "lon": 23.20227}, {"lat": 48.99319, "lon": 23.62867}, {
-  "lat": 50.47892,
-  "lon": 24.27876
-}, {"lat": 49.44912, "lon": 23.00434}, {"lat": 49.25634, "lon": 23.84771}, {
-  "lat": 49.15712,
-  "lon": 23.04949
-}, {"lat": 49.01808, "lon": 22.93619}, {"lat": 47.89696, "lon": 31.31945}, {
-  "lat": 47.39691,
-  "lon": 32.44898
-}, {"lat": 46.85583, "lon": 31.38605}, {"lat": 47.24396, "lon": 32.9904}, {
-  "lat": 47.86451,
-  "lon": 31.57299
-}, {"lat": 47.35554, "lon": 31.23568}, {"lat": 47.62024, "lon": 31.53282}, {
-  "lat": 47.85988,
-  "lon": 30.58594
-}, {"lat": 47.63071, "lon": 30.98522}, {"lat": 47.69731, "lon": 31.85022}, {
-  "lat": 47.82005,
-  "lon": 32.83195
-}, {"lat": 47.94657, "lon": 30.34527}, {"lat": 46.95883, "lon": 32.04574}, {
-  "lat": 47.69302,
-  "lon": 32.51146
-}, {"lat": 47.30095, "lon": 31.78431}, {"lat": 46.62872, "lon": 31.55788}, {
-  "lat": 48.04956,
-  "lon": 30.84618
-}, {"lat": 47.08723, "lon": 32.81668}, {"lat": 47.72118, "lon": 29.95937}, {
-  "lat": 45.98654,
-  "lon": 29.42637
-}, {"lat": 47.93592, "lon": 29.62034}, {"lat": 46.18137, "lon": 30.31677}, {
-  "lat": 46.48247,
-  "lon": 30.20107
-}, {"lat": 47.19761, "lon": 30.90712}, {"lat": 45.69437, "lon": 28.60256}, {
-  "lat": 47.0783,
-  "lon": 29.849
-}, {"lat": 47.18611, "lon": 29.94392}, {"lat": 45.3471, "lon": 28.83551}, {
-  "lat": 45.46228,
-  "lon": 29.25728
-}, {"lat": 48.09162, "lon": 29.11051}, {"lat": 46.82459, "lon": 30.95467}, {
-  "lat": 47.76066,
-  "lon": 29.52799
-}, {"lat": 47.53956, "lon": 29.4595}, {"lat": 47.84494, "lon": 30.26648}, {
-  "lat": 45.62176,
-  "lon": 29.47203
-}, {"lat": 46.24472, "lon": 30.44071}, {"lat": 46.85377, "lon": 30.0507}, {
-  "lat": 45.46074,
-  "lon": 28.28619
-}, {"lat": 48.1323, "lon": 30.07954}, {"lat": 46.02001, "lon": 29.68472}, {
-  "lat": 46.18669,
-  "lon": 29.1481
-}, {"lat": 45.84097, "lon": 29.61416}, {"lat": 47.33442, "lon": 29.75012}, {
-  "lat": 47.38455,
-  "lon": 30.18974
-}, {"lat": 49.78918, "lon": 33.72974}, {"lat": 50.36631, "lon": 33.98483}, {
-  "lat": 49.40328,
-  "lon": 33.21939
-}, {"lat": 50.1186, "lon": 32.44537}, {"lat": 49.8266, "lon": 34.52556}, {
-  "lat": 50.21026,
-  "lon": 34.35768
-}, {"lat": 49.4541, "lon": 35.13685}, {"lat": 49.14974, "lon": 34.2049}, {
-  "lat": 49.21257,
-  "lon": 33.83926
-}, {"lat": 50.06367, "lon": 34.74958}, {"lat": 49.08949, "lon": 33.42899}, {
-  "lat": 50.35927,
-  "lon": 33.27192
-}, {"lat": 50.01182, "lon": 33.00053}, {"lat": 49.44448, "lon": 34.84125}, {
-  "lat": 49.9653,
-  "lon": 33.60975
-}, {"lat": 49.33617, "lon": 34.31734}, {"lat": 49.78643, "lon": 32.69531}, {
-  "lat": 50.24443,
-  "lon": 32.50957
-}, {"lat": 49.57546, "lon": 34.53741}, {"lat": 49.56482, "lon": 34.07993}, {
-  "lat": 49.5952,
-  "lon": 33.17991
-}, {"lat": 49.7842, "lon": 33.29012}, {"lat": 50.26794, "lon": 32.93049}, {
-  "lat": 49.71434,
-  "lon": 35.15865
-}, {"lat": 49.85407, "lon": 33.30162}, {"lat": 50.99853, "lon": 26.74072}, {
-  "lat": 51.41825,
-  "lon": 26.14008
-}, {"lat": 50.59633, "lon": 26.67051}, {"lat": 50.43068, "lon": 25.34082}, {
-  "lat": 50.40733,
-  "lon": 25.76208
-}, {"lat": 51.57034, "lon": 26.56614}, {"lat": 50.25318, "lon": 25.47283}, {
-  "lat": 50.52904,
-  "lon": 26.25509
-}, {"lat": 50.61899, "lon": 27.16284}, {"lat": 50.87923, "lon": 26.45474}, {
-  "lat": 50.51239,
-  "lon": 25.6069
-}, {"lat": 50.32888, "lon": 26.52117}, {"lat": 50.61916, "lon": 26.25235}, {
-  "lat": 50.9242,
-  "lon": 26.49525
-}, {"lat": 51.31782, "lon": 26.62279}, {"lat": 50.13061, "lon": 25.25019}, {
-  "lat": 50.30931,
-  "lon": 34.89773
-}, {"lat": 51.14582, "lon": 34.31219}, {"lat": 51.19594, "lon": 33.82742}, {
-  "lat": 50.4257,
-  "lon": 35.48412
-}, {"lat": 51.67831, "lon": 33.9093}, {"lat": 51.24779, "lon": 33.22008}, {
-  "lat": 50.78104,
-  "lon": 35.2577
-}, {"lat": 51.55472, "lon": 33.38796}, {"lat": 50.58157, "lon": 34.47321}, {
-  "lat": 50.5553,
-  "lon": 33.80115
-}, {"lat": 50.83563, "lon": 33.87806}, {"lat": 51.33379, "lon": 33.86467}, {
-  "lat": 50.74894,
-  "lon": 33.47122
-}, {"lat": 52.18952, "lon": 34.02946}, {"lat": 50.90773, "lon": 34.79851}, {
-  "lat": 50.48647,
-  "lon": 34.96588
-}, {"lat": 51.86491, "lon": 33.47277}, {"lat": 51.94679, "lon": 33.78502}, {
-  "lat": 49.44775,
-  "lon": 24.93691
-}, {"lat": 48.80367, "lon": 26.03434}, {"lat": 49.06408, "lon": 25.38752}, {
-  "lat": 49.07129,
-  "lon": 26.22282
-}, {"lat": 48.63802, "lon": 25.73341}, {"lat": 49.66249, "lon": 25.77976}, {
-  "lat": 49.66335,
-  "lon": 25.14273
-}, {"lat": 49.43333, "lon": 25.14856}, {"lat": 50.09354, "lon": 25.72929}, {
-  "lat": 48.84659,
-  "lon": 25.98559
-}, {"lat": 49.09344, "lon": 25.17483}, {"lat": 49.52911, "lon": 26.13939}, {
-  "lat": 49.27145,
-  "lon": 25.13483
-}, {"lat": 49.30063, "lon": 25.68895}, {"lat": 49.5516, "lon": 25.60244}, {
-  "lat": 49.00537,
-  "lon": 25.79659
-}, {"lat": 50.12375, "lon": 26.1145}, {"lat": 49.45495, "lon": 36.84608}, {
-  "lat": 48.91268,
-  "lon": 37.0253
-}, {"lat": 48.85363, "lon": 36.55186}, {"lat": 50.161, "lon": 35.52378}, {
-  "lat": 49.90162,
-  "lon": 36.35153
-}, {"lat": 49.83759, "lon": 35.6139}, {"lat": 50.05474, "lon": 37.38699}, {
-  "lat": 50.28751,
-  "lon": 36.9441
-}, {"lat": 49.68086, "lon": 36.35462}, {"lat": 49.85218, "lon": 37.7334}, {
-  "lat": 50.10795,
-  "lon": 36.12116
-}, {"lat": 49.19626, "lon": 35.2383}, {"lat": 50.27532, "lon": 35.96975}, {
-  "lat": 49.19128,
-  "lon": 37.26768
-}, {"lat": 49.29308, "lon": 35.75329}, {"lat": 49.84342, "lon": 35.31332}, {
-  "lat": 49.37324,
-  "lon": 35.44876
-}, {"lat": 50.06281, "lon": 35.16552}, {"lat": 49.7097, "lon": 37.61204}, {
-  "lat": 49.14665,
-  "lon": 36.32217
-}, {"lat": 49.72086, "lon": 35.86126}, {"lat": 49.37376, "lon": 36.22124}, {
-  "lat": 49.86866,
-  "lon": 36.93844
-}, {"lat": 49.15352, "lon": 35.86676}, {"lat": 49.98934, "lon": 36.22982}, {
-  "lat": 49.84308,
-  "lon": 36.68404
-}, {"lat": 49.99569, "lon": 35.25873}, {"lat": 46.63061, "lon": 32.44554}, {
-  "lat": 46.84107,
-  "lon": 33.43397
-}, {"lat": 47.32223, "lon": 33.29887}, {"lat": 47.15383, "lon": 33.94689}, {
-  "lat": 47.25597,
-  "lon": 34.33399
-}, {"lat": 47.48909, "lon": 33.53165}, {"lat": 46.16781, "lon": 34.79061}, {
-  "lat": 46.52504,
-  "lon": 32.55558
-}, {"lat": 46.518, "lon": 34.2897}, {"lat": 46.38822, "lon": 32.10548}, {
-  "lat": 46.25502,
-  "lon": 33.29338
-}, {"lat": 46.81274, "lon": 33.50469}, {"lat": 46.84879, "lon": 34.37141}, {
-  "lat": 47.50059,
-  "lon": 33.91205
-}, {"lat": 46.35115, "lon": 34.33365}, {"lat": 46.1206, "lon": 32.9104}, {
-  "lat": 46.62495,
-  "lon": 32.72226
-}, {"lat": 46.35424, "lon": 33.54813}, {"lat": 49.99826, "lon": 26.42899}, {
-  "lat": 49.03576,
-  "lon": 27.23236
-}, {"lat": 49.53649, "lon": 26.19724}, {"lat": 49.17137, "lon": 26.56408}, {
-  "lat": 49.26853,
-  "lon": 27.42256
-}, {"lat": 48.88658, "lon": 26.8475}, {"lat": 50.11791, "lon": 26.82158}, {
-  "lat": 48.6636,
-  "lon": 26.58091
-}, {"lat": 49.65322, "lon": 26.96938}, {"lat": 49.38475, "lon": 27.63165}, {
-  "lat": 48.83646,
-  "lon": 27.26996
-}, {"lat": 50.12117, "lon": 27.5029}, {"lat": 49.05636, "lon": 27.39355}, {
-  "lat": 49.76154,
-  "lon": 27.22052
-}, {"lat": 49.59743, "lon": 27.63147}, {"lat": 49.84497, "lon": 26.41422}, {
-  "lat": 49.4165,
-  "lon": 26.97435
-}, {"lat": 49.01499, "lon": 26.33852}, {"lat": 50.16769, "lon": 27.05607}, {
-  "lat": 49.19266,
-  "lon": 26.83703
-}, {"lat": 49.28947, "lon": 31.44528}, {"lat": 49.95432, "lon": 32.14325}, {
-  "lat": 49.24639,
-  "lon": 30.11164
-}, {"lat": 49.0749, "lon": 30.96634}, {"lat": 49.66902, "lon": 32.04214}, {
-  "lat": 49.04159,
-  "lon": 32.10153
-}, {"lat": 49.75468, "lon": 31.46124}, {"lat": 48.94203, "lon": 30.98539}, {
-  "lat": 49.41582,
-  "lon": 31.25628
-}, {"lat": 49.25377, "lon": 30.82729}, {"lat": 48.96383, "lon": 30.33428}, {
-  "lat": 48.98975,
-  "lon": 29.82599
-}, {"lat": 49.22768, "lon": 31.85143}, {"lat": 48.88967, "lon": 30.68911}, {
-  "lat": 48.72454,
-  "lon": 30.19884
-}, {"lat": 48.80865, "lon": 29.98839}, {"lat": 49.46594, "lon": 32.02223}, {
-  "lat": 49.67829,
-  "lon": 32.30856
-}, {"lat": 49.07627, "lon": 32.66716}, {"lat": 49.01327, "lon": 31.39412}, {
-  "lat": 48.24217,
-  "lon": 25.17946
-}, {"lat": 48.1493, "lon": 26.26196}, {"lat": 48.08578, "lon": 25.92894}, {
-  "lat": 48.52198,
-  "lon": 25.84791
-}, {"lat": 48.46876, "lon": 26.82346}, {"lat": 48.44284, "lon": 25.76071}, {
-  "lat": 48.21865,
-  "lon": 26.27415
-}, {"lat": 47.99017, "lon": 25.08642}, {"lat": 48.44456, "lon": 27.40814}, {
-  "lat": 48.16114,
-  "lon": 25.72243
-}, {"lat": 48.50447, "lon": 26.4913}, {"lat": 51.18032, "lon": 32.81444}, {
-  "lat": 50.74774,
-  "lon": 31.39189
-}, {"lat": 51.25397, "lon": 32.427}, {"lat": 50.49917, "lon": 32.72329}, {
-  "lat": 50.75958,
-  "lon": 32.52794
-}, {"lat": 50.86155, "lon": 32.39044}, {"lat": 50.9539, "lon": 30.88017}, {
-  "lat": 51.56708,
-  "lon": 32.95195
-}, {"lat": 51.77462, "lon": 32.2514}, {"lat": 50.97124, "lon": 32.16505}, {
-  "lat": 51.52262,
-  "lon": 32.2162
-}, {"lat": 51.04866, "lon": 31.88799}, {"lat": 52.00413, "lon": 33.27793}, {
-  "lat": 50.93605,
-  "lon": 31.56733
-}, {"lat": 50.51325, "lon": 32.58802}, {"lat": 51.80363, "lon": 31.09285}, {
-  "lat": 51.46854,
-  "lon": 31.91717
-}, {"lat": 51.52227, "lon": 32.50168}, {"lat": 50.66294, "lon": 32.91933}, {
-  "lat": 50.95871,
-  "lon": 31.92507
-}, {"lat": 51.49807, "lon": 31.29009}, {"lat": 51.82388, "lon": 31.94258}]
