--- src/api.js
@@ -1,6 +1,0 @@
-export class Point {
-  constructor(x, y) {
-    this.x = x;
-    this.y = y;
-  }
-}
