--- src/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Mapper</title>
-</head>
-<body>
-<div class="mapping">
-  <div class="map">
-    <canvas id="map" width="982" height="673"></canvas>
-  </div>
-  <div class="properties">
-    <div class="debug">
-      <canvas class="debug__canvas" width="100" height="100"></canvas>
-    </div>
-    <label for="region">Region</label>
-    <select name="region" id="region" class="properties__region"></select>
-    <label for="district">District</label>
-    <select name="district" id="district" class="properties__district"></select>
-    <button class="properties__save-button">Save</button>
-  </div>
-</div>
-<div class="progress">
-  <svg id="progress" width="982" height="673"></svg>
-</div>
-</body>
-</html>
