+++ src/sass/editor.scss
@@ -1,0 +1,20 @@
+html, body {
+  margin: 0;
+  body: 0;
+}
+
+.mapping {
+  display: flex;
+
+  .debug {
+    display: none;
+  }
+
+  .map {
+    flex-grow: 2;
+  }
+
+  .properties {
+    min-width: 480px;
+  }
+}
